+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/components/molecules/UserProfile.kt
@@ -6,34 +6,56 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import bisqapps.shared.presentation.generated.resources.Res
 import bisqapps.shared.presentation.generated.resources.img_bot_image
+import kotlinx.coroutines.flow.StateFlow
 import network.bisq.mobile.domain.data.replicated.user.profile.UserProfileVO
+import network.bisq.mobile.domain.data.replicated.user.reputation.ReputationScoreVO
 import network.bisq.mobile.presentation.ui.components.atoms.BisqText
 import network.bisq.mobile.presentation.ui.components.atoms.StarRating
+import network.bisq.mobile.presentation.ui.components.atoms.icons.LanguageIcon
+import network.bisq.mobile.presentation.ui.components.atoms.layout.BisqGap
+import network.bisq.mobile.presentation.ui.theme.BisqUIConstants
 import org.jetbrains.compose.resources.painterResource
 
-// TODO: Get params and render apt
 @Composable
-fun UserProfile(item: UserProfileVO, showUserName: Boolean = true) {
-    val fiveSystemScore: Double = 3.5 // TODO: item.reputationScore.fiveSystemScore
+fun UserProfile(
+    user: UserProfileVO,
+    reputation: StateFlow<ReputationScoreVO>,
+    supportedLanguageCodes: List<String>,
+    showUserName: Boolean = true,
+    modifier: Modifier = Modifier
+) {
+    val fiveSystemScore: Double = reputation.collectAsState().value.fiveSystemScore
 
-    Row(horizontalArrangement = Arrangement.spacedBy(12.dp), verticalAlignment = Alignment.CenterVertically) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier
+    ) {
         Image(
             painterResource(Res.drawable.img_bot_image), "",
-            modifier = Modifier.size(36.dp)
+            modifier = Modifier.size(BisqUIConstants.ScreenPadding3X)
         )
-        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
+        BisqGap.V1()
+        Column() {
             if (showUserName) {
                 BisqText.baseRegular(
-                    text = item.userName,
+                    text = user.userName,
                     singleLine = true,
                 )
+                BisqGap.VQuarter()
             }
             StarRating(fiveSystemScore)
+        }
+        BisqGap.V2()
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            LanguageIcon()
+            BisqText.smallRegularGrey(text = " : ")
+            BisqText.smallRegular(text = supportedLanguageCodes.joinToString(", ").uppercase())
         }
     }
 }
